+++ .gitignore
@@ -13,3 +13,4 @@
 .vscode
 *.ipynb
 /result
+.cargo
