+++ src/cdb/db.rs
@@ -1,7 +1,6 @@
 use {
     super::{
         binary::{BinaryLinkIF, BinaryLinkList},
-        clause::{Clause, ClauseIF},
         ema::ProgressLBD,
         property,
         watch_cache::*,
@@ -24,14 +23,14 @@ use std::{fs::File, io::Write, path::Path};
 #[derive(Clone, Debug)]
 pub struct ClauseDB {
     /// container of clauses
-    pub(super) clause: Vec<Clause>,
+    pub(crate) clause: Vec<Clause>,
     /// hashed representation of binary clauses.
     ///## Note
     /// This means a biclause \[l0, l1\] is stored at bi_clause\[l0\] instead of bi_clause\[!l0\].
     ///
     binary_link: BinaryLinkDB,
     /// container of watch literals
-    pub(super) watch_cache: Vec<WatchCache>,
+    pub(crate) watch_cache: Vec<WatchCache>,
     /// collected free clause ids.
     freelist: Vec<ClauseId>,
     /// see unsat_certificate.rs
@@ -45,7 +44,7 @@ pub struct ClauseDB {
 
     // bi-clause completion
     bi_clause_completion_queue: Vec<Lit>,
-    pub(super) num_bi_clause_completion: usize,
+    pub(crate) num_bi_clause_completion: usize,
 
     //
     //## clause rewarding
@@ -63,30 +62,30 @@ pub struct ClauseDB {
     //
     /// a working buffer for LBD calculation
     lbd_temp: Vec<usize>,
-    pub(super) lbd: ProgressLBD,
+    pub(crate) lbd: ProgressLBD,
 
     //
     //## statistics
     //
     /// the number of active (not DEAD) clauses.
-    pub(super) num_clause: usize,
+    pub(crate) num_clause: usize,
     /// the number of binary clauses.
-    pub(super) num_bi_clause: usize,
+    pub(crate) num_bi_clause: usize,
     /// the number of binary learnt clauses.
-    pub(super) num_bi_learnt: usize,
+    pub(crate) num_bi_learnt: usize,
     /// the number of clauses which LBDs are 2.
-    pub(super) num_lbd2: usize,
+    pub(crate) num_lbd2: usize,
     /// the present number of learnt clauses.
-    pub(super) num_learnt: usize,
+    pub(crate) num_learnt: usize,
     /// the number of reductions.
-    pub(super) num_reduction: usize,
+    pub(crate) num_reduction: usize,
     /// the number of reregistration of a bi-clause
-    pub(super) num_reregistration: usize,
+    pub(crate) num_reregistration: usize,
     /// Literal Block Entanglement
     /// EMA of LBD of clauses used in conflict analysis (dependency graph)
-    pub(super) lb_entanglement: Ema2,
+    pub(crate) lb_entanglement: Ema2,
     /// cutoff value used in the last `reduce`
-    pub(super) reduction_threshold: f64,
+    pub(crate) reduction_threshold: f64,
 
     //
     //## incremental solving
@@ -355,6 +354,7 @@ impl ClauseDBIF for ClauseDB {
             // }
             // assert!(c.is_dead());
             c.flags = FlagClause::empty();
+            c.used = 0;
 
             #[cfg(feature = "clause_rewarding")]
             {
@@ -393,6 +393,7 @@ impl ClauseDBIF for ClauseDB {
             ..
         } = self;
         let c = &mut clause[NonZeroU32::get(cid.ordinal) as usize];
+        c.used = 0;
         #[cfg(feature = "clause_rewarding")]
         {
             c.timestamp = *tick;
@@ -400,7 +401,6 @@ impl ClauseDBIF for ClauseDB {
         let len2 = c.lits.len() == 2;
         if len2 {
             c.rank = 1;
-            c.rank_old = 1;
 
             #[cfg(feature = "bi_clause_completion")]
             if learnt {
@@ -412,7 +412,6 @@ impl ClauseDBIF for ClauseDB {
             }
         } else {
             c.update_lbd(asg, lbd_temp);
-            c.rank_old = c.rank;
         }
         self.lbd.update(c.rank);
         *num_clause += 1;
@@ -450,6 +449,7 @@ impl ClauseDBIF for ClauseDB {
             cid = cid_used;
             let c = &mut self[cid];
             c.flags = FlagClause::empty();
+            c.used = 0;
             std::mem::swap(&mut c.lits, vec);
             c.search_from = 2;
         } else {
@@ -472,6 +472,7 @@ impl ClauseDBIF for ClauseDB {
             ..
         } = self;
         let c = &mut clause[NonZeroU32::get(cid.ordinal) as usize];
+        c.used = 0;
 
         #[cfg(feature = "clause_rewarding")]
         {
@@ -481,10 +482,8 @@ impl ClauseDBIF for ClauseDB {
         let len2 = c.lits.len() == 2;
         if len2 {
             c.rank = 1;
-            c.rank_old = 1;
         } else {
             c.update_lbd(asg, lbd_temp);
-            c.rank_old = c.rank;
             c.turn_on(FlagClause::LEARNT);
         }
         let l0 = c.lits[0];
@@ -736,10 +735,10 @@ impl ClauseDBIF for ClauseDB {
             //## Case:2
             //
             let old_l0 = c.lit0();
-            let old_l1 = c.lit0();
+            let old_l1 = c.lit1();
             std::mem::swap(&mut c.lits, new_lits);
             let l0 = c.lit0();
-            let l1 = c.lit0();
+            let l1 = c.lit1();
             watch_cache[!old_l0].remove_watch(&cid);
             watch_cache[!old_l1].remove_watch(&cid);
             binary_link.add(l0, l1, cid);
@@ -760,10 +759,10 @@ impl ClauseDBIF for ClauseDB {
             //## Case:3
             //
             let old_l0 = c.lit0();
-            let old_l1 = c.lit0();
+            let old_l1 = c.lit1();
             std::mem::swap(&mut c.lits, new_lits);
             let l0 = c.lit0();
-            let l1 = c.lit0();
+            let l1 = c.lit1();
 
             if (l0 == old_l0 && l1 == old_l1) || (l0 == old_l1 && l1 == old_l0) {
             } else if l0 == old_l0 {
@@ -1063,13 +1062,16 @@ impl ClauseDBIF for ClauseDB {
                 continue;
             }
             alives += 1;
+            if c.rank <= 4 {
+                continue;
+            }
             match setting {
                 ReductionType::RASonADD(_) => {
                     perm.push(OrderedProxy::new(i, c.reverse_activity_sum(asg)));
                 }
                 ReductionType::RASonALL(cutoff, _) => {
                     let value = c.reverse_activity_sum(asg);
-                    if cutoff < value.min(c.rank_old as f64) {
+                    if cutoff < value.min(c.rank as f64) {
                         perm.push(OrderedProxy::new(i, value));
                     }
                 }
@@ -1077,7 +1079,7 @@ impl ClauseDBIF for ClauseDB {
                     perm.push(OrderedProxy::new(i, c.lbd()));
                 }
                 ReductionType::LBDonALL(cutoff, _) => {
-                    let value = c.rank.min(c.rank_old);
+                    let value = c.rank;
                     if cutoff < value {
                         perm.push(OrderedProxy::new(i, value as f64));
                     }
@@ -1181,14 +1183,6 @@ impl ClauseDBIF for ClauseDB {
         while let Some(lit) = self.bi_clause_completion_queue.pop() {
             self.complete_bi_clauses_with(asg, lit);
         }
-    }
-    #[cfg(feature = "incremental_solver")]
-    fn make_permanent_immortal(&mut self, cid: ClauseId) {
-        self.eliminated_permanent.push(
-            self.clause[NonZeroU32::get(cid.ordinal) as usize]
-                .lits
-                .clone(),
-        );
     }
     #[cfg(feature = "boundary_check")]
     fn watch_cache_contains(&self, lit: Lit, cid: ClauseId) -> bool {
