--- src/primitive/mod.rs
@@ -1,6 +1,0 @@
-/// methods on clause activity
-pub mod ema;
-/// methods on binary link, namely binary clause
-pub mod luby;
-
-pub use self::{ema::*, luby::*};
