+++ src/solver/conflict.rs
@@ -2,6 +2,8 @@
 
 #[cfg(feature = "boundary_check")]
 use crate::assign::DebugReportIF;
+#[cfg(feature = "trail_saving")]
+use crate::assign::TrailSavingIF;
 
 use {
     super::State,
@@ -10,11 +12,12 @@ use {
         cdb::{ClauseDB, ClauseDBIF},
         types::*,
     },
+    std::collections::HashSet,
 };
 
 /// returns:
-/// - 0: if a new assigngment is generated by conflict analysis.
-/// - 1: if a binary link generated
+/// - 0: if a new assignment is generated by conflict analysis.
+/// - 1: if a binary link is generated
 /// - otherwise: it's the LBD of the learnt clause.
 #[allow(clippy::cognitive_complexity)]
 pub fn handle_conflict(
@@ -23,8 +26,11 @@ pub fn handle_conflict(
     state: &mut State,
     cc: &ConflictContext,
 ) -> Result<u16, SolverError> {
+    // `conflicting_level` should be calculated from cc.1 instead of cc.0.
+    // Because the conflicting_literal has two values assigned at different levels.
+    // We need larger one.
     #[cfg(feature = "chrono_BT")]
-    let mut conflicting_level = asg.decision_level();
+    let conflicting_level; // for now
     #[cfg(not(feature = "chrono_BT"))]
     let conflicting_level = asg.decision_level();
 
@@ -39,44 +45,38 @@ pub fn handle_conflict(
         }
     }
 
-    // If we can settle this conflict w/o restart, solver will get a big progress.
     #[cfg(feature = "chrono_BT")]
-    let chronobt = 1000 < asg.num_conflict && 0 < state.config.c_cbt_thr;
-    #[cfg(not(feature = "chrono_BT"))]
-    let chronobt: bool = false;
-
-    #[cfg(feature = "chrono_BT")]
-    {
-        let c = match cc.1 {
-            AssignReason::Implication(cid) => &cdb[cid],
-            _ => panic!(),
-        };
-        let max_level = c.iter().map(|l| asg.level(l.vi())).max().unwrap();
-
-        if chronobt
-            && state.config.c_cbt_thr < conflicting_level
-            && 1 == c.iter().filter(|l| level[l.vi()] == max_level).count()
-        {
-            if let Some(second_level) = c
+    match cc.1 {
+        AssignReason::BinaryLink(l) => {
+            conflicting_level = asg.var(l.vi()).level.max(asg.var(cc.0.vi()).level);
+        }
+        AssignReason::Implication(cid) => {
+            let c = &cdb[cid];
+            conflicting_level = c.iter().map(|l| asg.level(l.vi())).max().unwrap();
+            if 1 == c
                 .iter()
-                .map(|l| level[l.vi()])
-                .filter(|l| *l < max_level)
-                .max()
+                .filter(|l| asg.level(l.vi()) == conflicting_level)
+                .count()
             {
-                debug_assert!(0 < second_level);
-                asg.cancel_until(second_level);
-                return Ok(());
+                if let Some(second_level) = c
+                    .iter()
+                    .map(|l| asg.level(l.vi()))
+                    .filter(|l| *l < conflicting_level)
+                    .max()
+                {
+                    debug_assert!(0 < second_level);
+                    asg.cancel_until(second_level);
+                    return Ok(c.rank);
+                }
             }
         }
-        if max_level < conflicting_level {
-            conflicting_level = max_level;
-            asg.cancel_until(conflicting_level);
+        _ => {
+            panic!();
         }
     }
-
+    asg.cancel_until(conflicting_level);
     asg.handle(SolverEvent::Conflict);
 
-    state.derive20.clear();
     let assign_level = conflict_analyze(asg, cdb, state, cc).max(asg.root_level());
     let new_learnt = &mut state.new_learnt;
     let learnt_len = new_learnt.len();
@@ -99,7 +99,7 @@ pub fn handle_conflict(
         //
         match asg.assigned(l0) {
             Some(true) if asg.root_level() < asg.level(l0.vi()) => {
-                panic!("double assignment occured");
+                panic!("double assignment occurred");
                 // asg.lift_to_asserted(l0.vi());
             }
             Some(false) if asg.level(l0.vi()) == asg.root_level() => {
@@ -160,16 +160,50 @@ pub fn handle_conflict(
             AssignReason::None => unreachable!("handle_conflict"),
         }
     }
-    if chronobt && assign_level + state.config.c_cbt_thr <= conflicting_level {
-        asg.cancel_until(conflicting_level - 1);
+
+    // learnt clause quality based backtrack strategy switching
+    // Idea: If the learned clause is low quality, don’t trust it to justify a large backjump; use CBT/limited-backjump instead.
+    let bt_drift: Option<bool> = if cfg!(feature = "chrono_BT")
+        && assign_level > 0
+        && asg
+            .len_upto(conflicting_level)
+            .saturating_sub(asg.len_upto(assign_level))
+            >= 40
+    {
+        Some(true)
+    } else if cfg!(feature = "BT_drift")
+        && assign_level > 0
+        && new_learnt
+            .iter()
+            .map(|l| asg.level(l.vi()))
+            .collect::<HashSet<_>>()
+            .len() as f64
+            >= 2.5 * cdb.lbd.get().max(3.0)
+    {
+        Some(false)
     } else {
-        asg.cancel_until(assign_level);
+        None
+    };
+    asg.cancel_until(match bt_drift {
+        None => assign_level,
+        Some(false) => assign_level - 1,
+        Some(true) => conflicting_level - 1,
+    });
+    if bt_drift.is_some() {
+        #[cfg(feature = "trail_saving")]
+        {
+            asg.clear_saved_trail();
+        }
+        #[cfg(feature = "chrono_BT")]
+        {
+            state.num_chrono_bt += 1;
+        }
     }
-    debug_assert_eq!(asg.assigned(l0), None);
-    debug_assert_eq!(
-        new_learnt.iter().skip(1).map(|l| asg.level(l.vi())).max(),
-        Some(assign_level)
-    );
+    // debug_assert_eq!(asg.assigned(l0), None);
+    // debug_assert_eq!(
+    //     new_learnt.iter().skip(1).map(|l| asg.level(l.vi())).max(),
+    //     Some(assign_level)
+    // );
     let rank: u16;
     match cdb.new_clause(asg, new_learnt, true) {
         RefClause::Clause(cid) if learnt_len == 2 => {
@@ -178,18 +212,12 @@ pub fn handle_conflict(
 
             debug_assert_eq!(l0, cdb[cid].lit0());
             debug_assert_eq!(l1, cdb[cid].lit1());
-            debug_assert_eq!(asg.assigned(l1), Some(false));
             debug_assert_eq!(asg.assigned(l0), None);
+            debug_assert!(bt_drift.is_some() || asg.assigned(l1) == Some(false));
 
-            asg.assign_by_implication(
-                l0,
-                AssignReason::BinaryLink(!l1),
-                #[cfg(feature = "chrono_BT")]
-                assign_level,
-            );
-            // || check_graph(asg, cdb, l0, "biclause");
-            for cid in &state.derive20 {
-                cdb[cid].turn_on(FlagClause::DERIVE20);
+            if bt_drift.is_none() {
+                asg.assign_by_implication(l0, AssignReason::BinaryLink(!l1), assign_level);
+                cdb[cid].used = cdb[cid].used.saturating_add(1);
             }
             rank = 1;
             #[cfg(feature = "bi_clause_completion")]
@@ -200,20 +228,11 @@ pub fn handle_conflict(
             cdb[cid].set_birth(asg.num_conflict);
 
             debug_assert_eq!(cdb[cid].lit0(), l0);
-            debug_assert_eq!(asg.assigned(l0), None);
-            asg.assign_by_implication(
-                l0,
-                AssignReason::Implication(cid),
-                #[cfg(feature = "chrono_BT")]
-                assign_level,
-            );
-            // || check_graph(asg, cdb, l0, "clause");
-            rank = cdb[cid].rank;
-            if rank <= 20 {
-                for cid in &state.derive20 {
-                    cdb[cid].turn_on(FlagClause::DERIVE20);
-                }
+            if bt_drift.is_none_or(|up1| up1 && cdb[cid].is_unit_under(&*asg)) {
+                asg.assign_by_implication(l0, AssignReason::Implication(cid), assign_level);
+                cdb[cid].used = cdb[cid].used.saturating_add(1);
             }
+            rank = cdb[cid].rank;
         }
         RefClause::RegisteredClause(cid) => {
             debug_assert_eq!(learnt_len, 2);
@@ -221,19 +240,25 @@ pub fn handle_conflict(
                 (l0 == cdb[cid].lit0() && l1 == cdb[cid].lit1())
                     || (l0 == cdb[cid].lit1() && l1 == cdb[cid].lit0())
             );
-            debug_assert_eq!(asg.assigned(l1), Some(false));
             debug_assert_eq!(asg.assigned(l0), None);
+            debug_assert!(bt_drift.is_some() || asg.assigned(l1) == Some(false));
+            // if bt_drift.is_none() && asg.assigned(l1) != Some(false) {
+            //     dbg!(cc);
+            //     dbg!(asg.decision_level());
+            //     dbg!(cid);
+            //     for l in cdb[cid].iter() {
+            //         println!("{:?} @ {} => {:?}", l, asg.level(l.vi()), asg.assigned(*l));
+            //     }
+            //     panic!("here we are!");
+            // }
             rank = 1;
-            asg.assign_by_implication(
-                l0,
-                AssignReason::BinaryLink(!l1),
-                #[cfg(feature = "chrono_BT")]
-                assign_level,
-            );
-            // || check_graph(asg, cdb, l0, "registeredclause");
+            if bt_drift.is_none_or(|up1| up1 && cdb[cid].is_unit_under(&*asg)) {
+                asg.assign_by_implication(l0, AssignReason::BinaryLink(!l1), assign_level);
+                cdb[cid].used = cdb[cid].used.saturating_add(1);
+            }
         }
-        RefClause::Dead => unreachable!("handle_conflict::RefClause::Deaf"),
-        RefClause::EmptyClause => unreachable!("handel_conflict::RefClause::EmptyClause"),
+        RefClause::Dead => unreachable!("handle_conflict::RefClause::Dead"),
+        RefClause::EmptyClause => unreachable!("handle_conflict::RefClause::EmptyClause"),
         RefClause::UnitClause(_) => unreachable!("handle_conflict::RefClause::UnitClause"),
     }
     state.c_lvl.update(conflicting_level as f64);
@@ -241,7 +266,6 @@ pub fn handle_conflict(
     state
         .e_mode
         .update(conflicting_level as f64 - assign_level as f64);
-    state.derive20.clear();
     Ok(rank)
 }
 
@@ -258,10 +282,17 @@ fn conflict_analyze(
     learnt.clear();
     learnt.push(Lit::from(u32::MAX));
     let root_level = asg.root_level();
+    // assert_eq!(asg.decision_level(), asg.var(cc.0.vi()).level);
     let dl = asg.decision_level();
     let mut path_cnt = 0;
     let (mut p, mut reason) = cc;
-
+    // assert_eq!(
+    //     dl,
+    //     asg.var(p.vi()).level,
+    //     "conflict: {:?} occurred at a lower level than current dl",
+    //     cc
+    // );
+    /// trace and reward the var
     macro_rules! conflict_level {
         ($vi: expr) => {
             path_cnt += 1;
@@ -275,6 +306,7 @@ fn conflict_analyze(
             println!($($arg),*);
         };
     }
+    /// a logging facility
     macro_rules! trace_lit {
         ($lit: expr, $message: expr) => {
             #[cfg(feature = "trace_analysis")]
@@ -285,11 +317,15 @@ fn conflict_analyze(
             }
         };
     }
-
+    /// vi must be assigned at the current or lower decision level
     macro_rules! validate_vi {
         ($vi: expr) => {
             debug_assert!(!asg.var($vi).is(FlagVar::ELIMINATED));
-            debug_assert!(asg.assign($vi).is_some());
+            debug_assert!(
+                asg.assign($vi).is_some(),
+                "conflict_analyze(validate_vi) found unassigned: {:?}",
+                $vi
+            );
             debug_assert!(asg.level($vi) <= dl);
         };
     }
@@ -323,6 +359,14 @@ fn conflict_analyze(
     {
         trace_lit!("- handle conflicting literal", p);
         let vi = p.vi();
+        if asg.var(vi).assign.is_none() {
+            println!("L355: {cc:?}");
+            println!(
+                " - dl: {}, cc.0.level {}",
+                asg.decision_level(),
+                asg.var(cc.0.vi()).level
+            );
+        }
         validate_vi!(vi);
         set_seen!(vi);
         let lvl = asg.level(vi);
@@ -342,11 +386,29 @@ fn conflict_analyze(
                 let vi = l.vi();
                 if !asg.var(vi).is(FlagVar::CA_SEEN) {
                     validate_vi!(vi);
-                    debug_assert_eq!(asg.level(vi), dl, "strange level binary clause");
+                    // debug_assert_eq!(
+                    //     asg.level(vi),
+                    //     dl,
+                    //     "unblanced bi-clause {}-{} at conflict level {dl}, source: {}, cc: {:?}",
+                    //     p,
+                    //     l,
+                    //     asg.var(p.vi()).level,
+                    //     cc
+                    // );
                     // if root_level == asg.level(vi) { continue; }
                     set_seen!(vi);
                     trace_lit!(l, " - binary linked");
-                    conflict_level!(vi);
+                    let lvl = asg.level(vi);
+                    if dl == lvl {
+                        trace_lit!(l, " -- found another path");
+                        conflict_level!(vi);
+                    } else {
+                        trace_lit!(l, " -- push to earnt");
+                        // This path is created for chrono-BT.
+                        // So we must consider if `l` should be pushed or `!l`.
+                        // From debug log, it seems to be `!l`.
+                        learnt.push(!l);
+                    }
                 }
             }
             AssignReason::Implication(cid) => {
@@ -359,7 +421,7 @@ fn conflict_analyze(
                 debug_assert!(!cdb[cid].is_dead() && 2 < cdb[cid].len());
                 // if !cdb.update_at_analysis(asg, cid) {
                 if !cdb[cid].is(FlagClause::LEARNT) {
-                    state.derive20.push(cid);
+                    cdb[cid].used = cdb[cid].used.saturating_add(1);
                 }
                 if max_lbd < cdb[cid].rank {
                     max_lbd = cdb[cid].rank;
@@ -367,6 +429,26 @@ fn conflict_analyze(
                 }
                 for q in cdb[cid].iter().skip(1) {
                     let vi = q.vi();
+                    debug_assert!(asg.var(vi).assign.is_some());
+                    // if asg.var(vi).assign.is_none() {
+                    //     println!(
+                    //         "Implication clause contains non-assinged lit (cc: {:?}, at level: {:?})",
+                    //         cc,
+                    //         asg.var(cc.0.vi()).level
+                    //     );
+                    //     println!("p: {:?}, level: {:?}", p, asg.var(p.vi()).level);
+                    //     println!("dl: {}", asg.decision_level());
+                    //     println!(
+                    //         "c{:?}: {:?}",
+                    //         cid,
+                    //         cdb[cid]
+                    //             .iter()
+                    //             // .filter(|l| asg.var(l.vi()).assign.is_none())
+                    //             .map(|l| (l, asg.var(l.vi()).level))
+                    //             .collect::<Vec<_>>()
+                    //     );
+                    //     panic!();
+                    // }
                     validate_vi!(vi);
                     if !asg.var(vi).is(FlagVar::CA_SEEN) {
                         let lvl = asg.level(vi);
@@ -444,7 +526,7 @@ fn conflict_analyze(
     if let Some(cid) = cid_with_max_lbd {
         cdb.update_at_analysis(asg, cid);
     }
-    debug_assert!(learnt.iter().all(|l| *l != !p));
+    // debug_assert!(learnt.iter().all(|l| *l != !p));
     debug_assert_eq!(asg.level(p.vi()), dl);
     learnt[0] = !p;
     trace!(
