+++ src/types/idx.rs
@@ -1,12 +1,25 @@
-use {
-    super::ClauseId,
-    std::{fmt, num::NonZeroU32},
-};
+use std::{fmt, num::NonZeroU32};
+
+/// Decision Level Representation.
+pub type DecisionLevel = u32;
+
+/// 'Variable' identifier or 'variable' index, starting with one.
+/// Implementation note: NonZeroUsize can be used but requires a lot of changes.
+/// The current abstraction is incomplete.
+pub type VarId = usize;
 
 /// API for Clause Id.
 pub trait ClauseIdIF {
     /// return `true` if a given clause id is made from a `Lit`.
     fn is_lifted_lit(&self) -> bool;
+}
+
+/// Clause identifier, or clause index, starting with one.
+/// Note: ids are re-used after 'garbage collection'.
+#[derive(Clone, Copy, Eq, Hash, Ord, PartialEq, PartialOrd)]
+pub struct ClauseId {
+    /// a sequence number.
+    pub ordinal: NonZeroU32,
 }
 
 impl Default for ClauseId {
